requirements.txt
bcrypt==3.1.1
Django==1.9.7
django-pyodbc-azure==1.9.6.0
PyYAML==3.11
blinker==1.4
coverage==4.0.3
djangorestframework==3.1.3
mongoengine==0.9.0
nose==1.3.7
psycopg2==2.6.2
pymongo==3.2.1
pyodbc==4.0.17
python-dateutil==2.6.0
pytz==2016.1
teamcity-messages==1.17